fix(api): P2-5 Task 2 fix round 1/5 — explicit SET NULL column list, pin CHECK constraints to shared enums (#4192)
Composite ON DELETE SET NULL on ai_agent_op_evidence_run_org_fk and
ai_agent_graduation_intent_org_fk now names the explicit PG15 column list
(SET NULL (run_id) / SET NULL (promoted_intent_id)) — without it Postgres
nulls every FK column including the NOT NULL org_id, which would 23502-abort
the parent delete instead of preserving the evidence row as intended.
Documented the same invariant on the Drizzle mirrors, which cannot express
a column list via foreignKey(...).onDelete('set null').

Hoisted checkConstraintLiterals() into a shared
apps/api/src/db/schema/checkConstraintTestHelpers.ts and added four
assertions pinning ai_agent_op_evidence's three CHECK constraints and
ai_agent_graduation.state's CHECK against their @breeze/shared enums —
the file comment claiming this was already true "by construction" was
unenforced.
